build: warnings are errors, at a higher level
/W4 /WX on MSVC, -Wall -Wextra -Werror on gcc and clang. 86 warnings fixed
rather than suppressed; what is left is /wd4996 for the deprecated CRT names
and one amalgam artifact under two names, /wd4505 and -Wno-unused-function,
for a static helper only one of the concatenated files uses.

Most of it was mechanical: 44 unused parameters lose their names, 31 unused
locals and 5 write-only ones go. Worth naming individually:

- Arena::Push and Arena::PopTo each guarded against a null `this`. Every
  caller already null-checks the arena before it gets there, so the guards
  were dead and -Wnonnull-compare was right.
- Arena::PopTo's `pos`, `current` and `prev` shadowed the members of the very
  struct it walks; the loop now says `node` / `prevNode`.
- component/Label masked its text by writing 0xE2 over the buffer and then
  writing '*' over that. The first loop was dead and truncated a constant.
- component/DescriptionList packed rows through `El* pending[16]` that it only
  ever stored null into and never read — a counter wearing an array, and an
  unbounded one: nothing stopped count from running past 16.
- MurmurHash2's tail falls through on purpose; it says so now.
- kPi was defined five times, and the amalgam turned that into shadowing.
  One constant in Gpui.h.
- Six Listen() handlers narrowed intptr_t to int implicitly.

Two are left as they were, deliberately:
- story.cpp's ToolbarCheckRow ignores `plain`. StoryToolbarOpt documents it as
  "menu() rather than menu_with_check(): no check column", two callers pass
  it, and the check cell is drawn regardless. Marked with a TODO rather than
  changing how the story renders on the way past a warning.
- The amalgamator now preserves line numbers. Stripping comments, dropping
  internal includes and collapsing blank lines each shifted them, so every
  diagnostic pointed a few lines off the real one — which -Werror makes worth
  the blank lines it costs.

macOS gets the same flags but is still uncompiled; the build Mac stayed
unreachable.

+++ src/sys/SysInfo_win.cpp
@@ -65,7 +65,7 @@ static void RefreshMemory(SysState* s) {
     if (GlobalMemoryStatusEx(&ms)) {
         s->memTotal = ms.ullTotalPhys;
         s->memUsed = ms.ullTotalPhys - ms.ullAvailPhys;
-        s->mem = ms.dwMemoryLoad;
+        s->mem = (float)ms.dwMemoryLoad;
     }
 }
